tune parameters

+++ analysis/check.py
@@ -36,8 +36,8 @@ epoch = 10
 with torch.autograd.detect_anomaly():
 	for e in range(epoch):
 		i = random.randint(0, 3)
-		data  = torch.tensor(dataset[i])
-		label = torch.tensor(labelset[i])
+		data  = torch.tensor(dataset[i], dtype=torch.float)
+		label = torch.tensor(labelset[i], dtype=torch.float)
 
 		data = torch.autograd.Variable(data).cuda()
 		label = torch.autograd.Variable(label).cuda()
